Move Frontend `assets` Directory To Parent Directory, And Remove Redundant `react.svg`
