+++ nextstep-frontend/assets/defaultProfileImage.jpg
