+++ nextstep-frontend/src/pages/Profile.tsx
@@ -4,7 +4,7 @@ import { ExpandMore, ExpandLess } from '@mui/icons-material';
 import {getUserAuth} from "../handlers/userAuth.ts";
 import api from "../serverApi.ts";
 import {UserProfile} from "../models/UserProfile.ts";
-import defaultProfileImage from '../assets/defaultProfileImage.jpg';
+import defaultProfileImage from '../../assets/defaultProfileImage.jpg';
 
 const Profile: React.FC = () => {
   const [image, setImage] = useState<string | null>(null);
